+++ yde-landlords/src/main/java/com/xt/landlords/game/regular/phase/RegularClearPhaseData.java
@@ -2,6 +2,8 @@ package com.xt.landlords.game.regular.phase;
 
 import org.sunyata.octopus.model.PhaseData;
 
+import java.math.BigDecimal;
+
 /**
  * Created by leo on 17/5/16.
  */
@@ -9,16 +11,16 @@ public class RegularClearPhaseData extends PhaseData {
 
     private String serialNo;
 
-    public int getTotalMoney() {
+    public BigDecimal getTotalMoney() {
         return totalMoney;
     }
 
-    public RegularClearPhaseData setTotalMoney(int totalMoney) {
+    public RegularClearPhaseData setTotalMoney(BigDecimal totalMoney) {
         this.totalMoney = totalMoney;
         return this;
     }
 
-    private int totalMoney;//获得的总奖金
+    private BigDecimal totalMoney;//获得的总奖金
 
     public RegularClearPhaseData setSerialNo(String serialNo) {
         this.serialNo = serialNo;
